[test] Android Test: spilit robot methods
